requirements.txt
flask
flask-mysqldb
pycryptodome